requirements.txt
fastapi
uvicorn
httpx
streamlit
python-dotenv
langchain
langchain-google-genai
pyperclip
